refactor: modularize autonomous development tests
