+++ V2_COMPLIANCE_PROGRESS_TRACKER.md
@@ -169,6 +169,13 @@
 - **Completion Date**: 2025-08-24
 - **Summary**: Refactored from 722 to 55 lines by extracting manager, AI processor, coordinator, and config modules. New orchestrator coordinates components and maintains functionality.
 
+### MAJOR-005: Health Metrics Collector ✅
+- **File**: `src/core/health/metrics/collector.py`
+- **Status**: Completed
+- **Assigned To**: Victor Dixon
+- **Completion Date**: 2025-08-24
+- **Summary**: Consolidated metrics collection into other health monitoring modules, eliminating the monolithic collector file.
+
 ## 📋 AVAILABLE CONTRACTS FOR CLAIMING
 
 > **Note:** Remaining contract descriptions include current line counts for context only. There is no strict LOC target—deliver clean, production-ready, tested modules that honor SRP and SOLID principles.
